fixed various build warnings about aria labels etc, fixed vehicle photo blowing up on hover of driver pic; fixed some console warnings
